Move settings tray out of Panel
We are keeping this visually removed from the panel after all.
